requirements.txt
groq
python-dotenv
pandas
scikit-learn
jupyter
matplotlib
seaborn
datasets